requirements.txt
# Airflow core
 
apache-airflow-providers-databricks==3.4.0
apache-airflow-providers-http==4.3.0

# Integrações e funcionalidades adicionais
requests==2.30.0
slack-sdk==3.27.0
pandas==2.3.1
matplotlib==3.8.4
pyspark==3.5.1


# Utilidades e suporte
databricks-sql-connector==2.9.3
#3de1479a2be654c4
#dbc-8e283846-57c0.cloud.databricks.com
